--- a/benchmarks/create-pages/src/pages/index.js
+++ b/benchmarks/create-pages/src/pages/index.js
@@ -1,0 +1,3 @@
+import React from "react"
+
+export default () => <div>Hello world!</div>