--- a/benchmarks/create-pages/src/templates/blank.js
+++ b/benchmarks/create-pages/src/templates/blank.js
@@ -1,0 +1,3 @@
+import React from "react"
+
+export default () => <div>Yo!</div>