--- a/packages/gatsby/src/bootstrap/index.js
+++ b/packages/gatsby/src/bootstrap/index.js
@@ -7,6 +7,7 @@ const md5File = require(`md5-file/promise`)
 const crypto = require(`crypto`)
 const del = require(`del`)
 const path = require(`path`)
+const convertHrtime = require(`convert-hrtime`)
 
 const apiRunnerNode = require(`../utils/api-runner-node`)
 const { graphql } = require(`graphql`)
@@ -27,6 +28,7 @@ const {
 const {
   runInitialQueries,
 } = require(`../internal-plugins/query-runner/page-query-runner`)
+const queryQueue = require(`../internal-plugins/query-runner/query-queue`)
 const { writePages } = require(`../internal-plugins/query-runner/pages-writer`)
 const {
   writeRedirects,
@@ -174,7 +176,9 @@ module.exports = async (args: BootstrapArgs) => {
   const siteDir = `${program.directory}/.cache`
   const tryRequire = `${__dirname}/../utils/test-require-error.js`
   try {
-    await fs.copy(srcDir, siteDir, { clobber: true })
+    await fs.copy(srcDir, siteDir, {
+      clobber: true,
+    })
     await fs.copy(tryRequire, `${siteDir}/test-require-error.js`, {
       clobber: true,
     })
@@ -344,7 +348,17 @@ module.exports = async (args: BootstrapArgs) => {
   // Run queries
   activity = report.activityTimer(`run graphql queries`)
   activity.start()
-  await runInitialQueries()
+  const startQueries = process.hrtime()
+  queryQueue.on(`task_finish`, () => {
+    const stats = queryQueue.getStats()
+    // console.log({ stats })
+    activity.setStatus(
+      `${stats.total}/${stats.peak} ${(
+        stats.total / convertHrtime(process.hrtime(startQueries)).seconds
+      ).toFixed(2)} queries/second`
+    )
+  })
+  await runInitialQueries(activity)
   activity.end()
 
   // Write out files.
@@ -375,7 +389,9 @@ module.exports = async (args: BootstrapArgs) => {
       activity.start()
       apiRunnerNode(`onPostBootstrap`).then(() => {
         activity.end()
-        resolve({ graphqlRunner })
+        resolve({
+          graphqlRunner,
+        })
       })
     }
   }, 100)
@@ -391,7 +407,9 @@ module.exports = async (args: BootstrapArgs) => {
     report.info(`bootstrap finished - ${process.uptime()} s`)
     report.log(``)
     emitter.emit(`BOOTSTRAP_FINISHED`)
-    return { graphqlRunner }
+    return {
+      graphqlRunner,
+    }
   } else {
     return new Promise(resolve => {
       // Wait until all side effect jobs are finished.