--- a/packages/gatsby/src/bootstrap/load-plugins/load.js
+++ b/packages/gatsby/src/bootstrap/load-plugins/load.js
@@ -5,6 +5,7 @@ const path = require(`path`)
 const crypto = require(`crypto`)
 const glob = require(`glob`)
 const { store } = require(`../../redux`)
+const existsSync = require(`fs-exists-cached`).sync
 
 function createFileContentHash(root, globPattern) {
   const hash = crypto.createHash(`md5`)
@@ -34,12 +35,12 @@ function createFileContentHash(root, globPattern) {
  */
 function resolvePlugin(pluginName) {
   // Only find plugins when we're not given an absolute path
-  if (!fs.existsSync(pluginName)) {
+  if (!existsSync(pluginName)) {
     // Find the plugin in the local plugins folder
     const resolvedPath = slash(path.resolve(`./plugins/${pluginName}`))
 
-    if (fs.existsSync(resolvedPath)) {
-      if (fs.existsSync(`${resolvedPath}/package.json`)) {
+    if (existsSync(resolvedPath)) {
+      if (existsSync(`${resolvedPath}/package.json`)) {
         const packageJSON = JSON.parse(
           fs.readFileSync(`${resolvedPath}/package.json`, `utf-8`)
         )