--- a/packages/gatsby/src/internal-plugins/internal-data-bridge/gatsby-node.js
+++ b/packages/gatsby/src/internal-plugins/internal-data-bridge/gatsby-node.js
@@ -40,6 +40,7 @@ function transformPackageJson(json) {
   return json
 }
 
+const createPageId = path => `SitePage ${path}`
 exports.sourceNodes = ({ actions, store }) => {
   const { createNode } = actions
   const state = store.getState()
@@ -136,37 +137,3 @@ exports.sourceNodes = ({ actions, store }) => {
     }
   })
 }
-
-const createPageId = path => `SitePage ${path}`
-
-exports.onCreatePage = ({ page, actions }) => {
-  const { createNode } = actions
-  // eslint-disable-next-line
-  const { updatedAt, ...pageWithoutUpdated } = page
-
-  // Add page.
-  createNode({
-    ...pageWithoutUpdated,
-    id: createPageId(page.path),
-    parent: `SOURCE`,
-    children: [],
-    internal: {
-      type: `SitePage`,
-      contentDigest: crypto
-        .createHash(`md5`)
-        .update(JSON.stringify(page))
-        .digest(`hex`),
-      description:
-        page.pluginCreatorId === `Plugin default-site-plugin`
-          ? `Your site's "gatsby-node.js"`
-          : page.pluginCreatorId,
-    },
-  })
-}
-
-// Listen for DELETE_PAGE and delete page nodes.
-emitter.on(`DELETE_PAGE`, action => {
-  const nodeId = createPageId(action.payload.path)
-  const node = getNode(nodeId)
-  boundActionCreators.deleteNode({ node })
-})