--- a/packages/gatsby/src/internal-plugins/query-runner/gatsby-node.js
+++ b/packages/gatsby/src/internal-plugins/query-runner/gatsby-node.js
@@ -3,12 +3,14 @@ const { watchComponent } = require(`./query-watcher`)
 let components = {}
 
 exports.onCreatePage = ({ page, store }) => {
-  const component = store.getState().components[page.componentPath]
+  // In development, watch the component to detect query changes.
+  if (process.env.NODE_ENV !== `production`) {
+    const component = store.getState().components[page.componentPath]
 
-  if (components[component.componentPath]) {
-    return
+    if (components[component.componentPath]) {
+      return
+    }
+
+    watchComponent(component.componentPath)
   }
-
-  // Watch the component to detect query changes.
-  watchComponent(component.componentPath)
 }