--- a/packages/gatsby/src/internal-plugins/query-runner/gatsby-node.js
+++ b/packages/gatsby/src/internal-plugins/query-runner/gatsby-node.js
@@ -8,9 +8,6 @@ exports.onCreatePage = ({ page, store }) => {
     const component = store.getState().components.get(page.componentPath)
 
     if (!component) {
-      console.log({ components: store.getState().components })
-      console.log({ page })
-      console.log({ component })
       process.exit()
     }