--- a/packages/gatsby/src/internal-plugins/query-runner/pages-writer.js
+++ b/packages/gatsby/src/internal-plugins/query-runner/pages-writer.js
@@ -12,6 +12,7 @@ let lastHash = null
 const writePages = async () => {
   bootstrapFinished = true
   let { program, jsonDataPaths, pages } = store.getState()
+  pages = [...pages.values()]
 
   const pagesComponentDependencies = {}