Spent the afternoon and evening going through critical path for creating pages.

Currently a 5000 page site can build in ~37 seconds. A very significant
improvement over the current.

This does necessitate one breaking change namely creating nodes for each page. This frankly
was more of academic interest and for debugging purposes and given that it adds a
large amount of slowdowns and *should* be rarely if ever used in sites, we should
hopefully be finish letting it go. In any case, if people _are_ using it, there's
far better ways of querying the same data.

## TODO
- [x] Calculate how much faster large sites are with this PR vs. current beta.
- [x] remove profile calls
- [x] Do node profiling on adding thousands of pages
- [x] Do memory profiling on same site.
- [x] Evaluate why "write out pages data" gets really slow at 25k
- [x] test freecodecamp
- [x] add benchmarking folder and first two sites — minimal createPages in memory and larger one that writes out markdown files
- [x] Add progress indicator for longer-running things like graphql queries + page building w/ final status at end
- [x] Remove extra packages from benchmark sites