@pieh 2018-06-29
> This does necessitate one breaking change namely creating nodes for each page. This frankly
> was more of academic interest and for debugging purposes and given that it adds a
> large amount of slowdowns and should be rarely if ever used in sites, we should
> hopefully be finish letting it go. In any case, if people are using it, there's
> far better ways of querying the same data.

Switching nodes reducer to map and mutating it instead of creating new state for every `CREATE_NODE` action will defenitely help here