@m-allanson 2018-06-29
> > Hmmm, gatsbyjs.org isn't building since StaticQueries aren't being run during builds. Is this a known issue? Been heads down last few days.

> I'm not seeing this here

Me neither. Seems like nice speedups for .org. Unscientific tests running `gatsby build` a few times (with warm cache):

```
current beta
✨  Done in 130.25s.
✨  Done in 97.00s.
✨  Done in 163.13s.

this PR
✨  Done in 72.71s.
✨  Done in 70.16s.
✨  Done in 74.66s.
```