@KyleAMathews 2018-06-30
Finally got gatsbyjs.org to work — had deleted public/static/d without also deleting the .cache directory 🤦‍♂️ 

Builds are taking anywhere between ~55-80 seconds on a warm cache. Not bad!