@KyleAMathews 2018-07-04
This is coming along!

Just built a 25k page site in a bit over 2 minutes

```
success open and validate gatsby-config — 0.082 s
success onPreBootstrap — 0.071 s
success delete html and css files from previous builds — 0.007 s
success copy gatsby files — 0.043 s
success source and transform nodes — 0.019 s
success building schema — 0.133 s
success createPages — 33.258 s
success createPagesStatefully — 0.184 s
success onPreExtractQueries — 0.000 s
success update schema — 0.056 s
success extract queries from components — 0.075 s
success run graphql queries — 21.329 s
success write out page data — 32.011 s
success write out redirect data — 0.001 s
success onPostBootstrap — 0.004 s

info bootstrap finished - 91.063 s

success Building production JavaScript and CSS bundles — 13.668 s
success Building static HTML for pages — 29.937 s
info Done building in 134.79 sec
```

Creating pages can still probably be made a lot faster. Writing out page data is getting weirdly slow w/ large number of pages but progress.