@zachgibson 2018-07-08
>This frankly was more of academic interest and for debugging purposes and given that it adds a large amount of slowdowns and should be rarely if ever used in sites, we should hopefully be finish letting it go.

Is this still the case, or will this be a viable solution? I’m looking at using Gatsby for a 400k page site, but not sure if things would just explode trying to do this.