@m-allanson 2018-07-09
@zachgibson the plan is to merge this in to Gatsby core as soon as we can!