@KyleAMathews 2018-07-09
@zachgibson did you see the initial sentence in that paragraph? What you quoted was referring to that breaking change — "This does necessitate one breaking change namely creating nodes for each page. "

This PR is definitely not experimental and will be merged soon, possibly today!