@KyleAMathews 2018-07-09
And on a 400k site — this will definitely be possible _eventually_ — even with this PR, Gatsby isn't quite ready to support that large of site. If you want to jump in and help with remaining issues on that, it will be possible to use Gatsby for 400k+ sites sooner than later.