@KyleAMathews 2018-07-09
@zachgibson they're not per se as the more immediate issue had been improving build speeds / memory usage. Those are getting solved so we'll be turning focus onto the next set of issues. Those are largely how to incrementally load site metadata. Currently we load all page metadata into the site which works well for sites under say 5k pages but the metadata starts getting overly large around there.