@KyleAMathews 2018-07-10
Charted memory usage when building a 10k markdown site. We can probably drop this quite a bit but acceptable for now.

<img width="593" alt="screen shot 2018-07-10 at 4 07 41 pm" src="https://user-images.githubusercontent.com/71047/42542357-722ea248-845b-11e8-8470-2e47d3212205.png">