@tradziej 2018-07-17
@KyleAMathews I don't know how this works but is there a need to keep all those empty directories (`public/static/d/*`) after the build (`gatsby build`) was finished?

https://github.com/gatsbyjs/gatsby/blob/d863bcb1ef555926e4ba5af313cb153cedad1291/packages/gatsby/src/bootstrap/index.js#L185