@KyleAMathews 2018-07-17
@tradziej yeah :-) I'm going to put up a PR in a sec to not pre-create the folders.