@KyleAMathews 2018-08-16
The benchmark sites https://github.com/gatsbyjs/gatsby/tree/master/benchmarks

Would love to hear more details about your site in another issue to see if we can figure out how to optimize it!